+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Response` class `contains_xml` method
 - `Prompt` class size warning
 - `Response` class size warning
+- `Session` class size warning
 ### Changed
 - `README.md` updated
 - `LLMModel` enum updated